+++ src/drivers/intel/dmic/CMakeLists.txt
@@ -1,3 +1,3 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
-add_local_sources(sof dmic.c)
+add_local_sources(sof dmic.c dmic_computed.c)
