+++ zephyr/CMakeLists.txt
@@ -154,6 +154,7 @@ if (CONFIG_SOC_SERIES_INTEL_CAVS_V15)
 
 	zephyr_library_sources_ifdef(CONFIG_INTEL_DMIC
 		${SOF_DRIVERS_PATH}/intel/dmic/dmic.c
+		${SOF_DRIVERS_PATH}/intel/dmic/dmic_computed.c
 	)
 
 	# Platform sources
@@ -207,6 +208,7 @@ if (CONFIG_SOC_SERIES_INTEL_CAVS_V18)
 
 	zephyr_library_sources_ifdef(CONFIG_INTEL_DMIC
 		${SOF_DRIVERS_PATH}/intel/dmic/dmic.c
+		${SOF_DRIVERS_PATH}/intel/dmic/dmic_computed.c
 	)
 
 	# Platform sources
@@ -262,6 +264,7 @@ if (CONFIG_SOC_SERIES_INTEL_CAVS_V20)
 
 	zephyr_library_sources_ifdef(CONFIG_INTEL_DMIC
 		${SOF_DRIVERS_PATH}/intel/dmic/dmic.c
+		${SOF_DRIVERS_PATH}/intel/dmic/dmic_computed.c
 	)
 
 	# Platform sources
@@ -319,6 +322,7 @@ if (CONFIG_SOC_SERIES_INTEL_CAVS_V25)
 
 	zephyr_library_sources_ifdef(CONFIG_INTEL_DMIC
 		${SOF_DRIVERS_PATH}/intel/dmic/dmic.c
+		${SOF_DRIVERS_PATH}/intel/dmic/dmic_computed.c
 	)
 
 	# Platform sources
